strip down inspector
